refactor: wrap ChatBubble with ChatMessageContainer to share more logic
