+++ lib/widgets/chat/chat_text_message.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:pass_emploi_app/presentation/chat_item.dart';
-import 'package:pass_emploi_app/ui/margins.dart';
 import 'package:pass_emploi_app/ui/text_styles.dart';
-import 'package:pass_emploi_app/widgets/chat/chat_bubble_widget.dart';
+import 'package:pass_emploi_app/widgets/chat/chat_message_container.dart';
 import 'package:pass_emploi_app/widgets/text_with_clickable_links.dart';
 
 class ChatTextMessage extends StatelessWidget {
@@ -14,23 +13,14 @@ class ChatTextMessage extends StatelessWidget {
   Widget build(BuildContext context) {
     final TextStyle textStyle =
         item is ConseillerTextMessageItem ? TextStyles.textSRegular() : TextStyles.textSRegular(color: Colors.white);
-    return Container(
-      margin: const EdgeInsets.only(top: 10, bottom: 10),
-      child: Column(
-        crossAxisAlignment: item is ConseillerTextMessageItem ? CrossAxisAlignment.start : CrossAxisAlignment.end,
-        children: [
-          ChatBubble(
-            isMyMessage: item is JeuneTextMessageItem,
-            child: SelectableTextWithClickableLinks(
-              item.content,
-              linkStyle: textStyle,
-              style: textStyle,
-            ),
-          ),
-          SizedBox(height: Margins.spacing_xs),
-          Text(item.caption, style: TextStyles.textXsRegular())
-        ],
+    return ChatMessageContainer(
+      content: SelectableTextWithClickableLinks(
+        item.content,
+        linkStyle: textStyle,
+        style: textStyle,
       ),
+      isMyMessage: item is JeuneTextMessageItem,
+      caption: item.caption,
     );
   }
 }
