+++ lib/widgets/chat/partage_message.dart
@@ -12,7 +12,7 @@ import 'package:pass_emploi_app/ui/drawables.dart';
 import 'package:pass_emploi_app/ui/margins.dart';
 import 'package:pass_emploi_app/ui/strings.dart';
 import 'package:pass_emploi_app/ui/text_styles.dart';
-import 'package:pass_emploi_app/widgets/chat/chat_bubble_widget.dart';
+import 'package:pass_emploi_app/widgets/chat/chat_message_container.dart';
 import 'package:pass_emploi_app/widgets/text_with_clickable_links.dart';
 
 class PartageMessage extends StatelessWidget {
@@ -22,30 +22,8 @@ class PartageMessage extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-      margin: const EdgeInsets.only(top: 10, bottom: 10),
-      child: Column(
-        crossAxisAlignment: item.sender == Sender.conseiller ? CrossAxisAlignment.start : CrossAxisAlignment.end,
-        children: [
-          _MessageBubble(item: item),
-          SizedBox(height: Margins.spacing_xs),
-          Text(item.caption, style: TextStyles.textXsRegular())
-        ],
-      ),
-    );
-  }
-}
-
-class _MessageBubble extends StatelessWidget {
-  final PartageMessageItem item;
-
-  _MessageBubble({required this.item});
-
-  @override
-  Widget build(BuildContext context) {
-    return ChatBubble(
-      isMyMessage: item.sender == Sender.jeune,
-      child: Column(
+    return ChatMessageContainer(
+      content: Column(
         crossAxisAlignment: CrossAxisAlignment.stretch,
         children: [
           _ContentMessage(content: item.content, sender: item.sender),
@@ -53,6 +31,8 @@ class _MessageBubble extends StatelessWidget {
           _PartageCard(item: item),
         ],
       ),
+      isMyMessage: item.sender == Sender.jeune,
+      caption: item.caption,
     );
   }
 }
